feat: change props for Product, add Category list

+++ src/Contracts/DataTransferObjects/Product/Product.cs
@@ -1,2 +1,2 @@
 ﻿namespace Contracts.DataTransferObjects;
-public record Product(int Id, string Name, List<string> Category, string Description, decimal Price);
+public record Product(int Id, string Name, List<string> Category, string Description, long Price);
